+++ modules/client/examples/02-tokenVoting-client/08-get-proposals.ts
@@ -13,7 +13,7 @@ import {
   TokenVotingClient,
   TokenVotingProposalListItem
 } from "@aragon/sdk-client";
-import { context } from "../00-setup/00-getting-started";
+import { context } from "../01-client/01-getting-started";
 
 // Create a plugin context from the Aragon SDK.
 const contextPlugin: ContextPlugin = ContextPlugin.fromContext(context);
@@ -34,60 +34,62 @@ console.log({ proposals });
 /* MARKDOWN
 Returns:
 
-```json
-[
-  {
-    id: "0x12345...",
-    dao: {
-      address: "0x1234567890123456789012345678901234567890",
-      name: "Cool DAO"
-    };
-    creatorAddress: "0x1234567890123456789012345678901234567890",
-    metadata: {
-      title: "Test Proposal",
-      summary: "Test Proposal Summary"
-    };
-    startDate: <Date>,
-    endDate: <Date>,
-    status: "Executed",
-    token: {
-      address: "0x1234567890123456789012345678901234567890,
-      name: "The Token",
-      symbol: "TOK",
-      decimals: 18
+```
+{ proposals:
+  [
+    {
+      id: "0x12345...",
+      dao: {
+        address: "0x1234567890123456789012345678901234567890",
+        name: "Cool DAO"
+      },
+      creatorAddress: "0x1234567890123456789012345678901234567890",
+      metadata: {
+        title: "Test Proposal",
+        summary: "Test Proposal Summary"
+      },
+      startDate: <Date>,
+      endDate: <Date>,
+      status: "Executed",
+      token: {
+        address: "0x1234567890123456789012345678901234567890,
+        name: "The Token",
+        symbol: "TOK",
+        decimals: 18
+      },
+      results {
+        yes: 100000n,
+        no: 77777n,
+        abstain: 0n
+      }
     },
-    results {
-      yes: 100000n,
-      no: 77777n,
-      abstain: 0n
+    {
+      id: "0x12345...",
+      dao: {
+        address: "0x1234567890123456789012345678901234567890",
+        name: "Cool DAO"
+      },
+      creatorAddress: "0x1234567890123456789012345678901234567890",
+      metadata: {
+        title: "Test Proposal 2",
+        summary: "Test Proposal Summary 2"
+      },
+      startDate: <Date>,
+      endDate: <Date>,
+      status: "Pending",
+      token: {
+        address: "0x1234567890123456789012345678901234567890,
+        name: "The Token",
+        symbol: "TOK",
+        decimals: 18
+      },
+      results {
+        yes: 100000n,
+        no: 77777n,
+        abstain: 0n
+      }
     }
-  },
-  {
-    id: "0x12345...",
-    dao: {
-      address: "0x1234567890123456789012345678901234567890",
-      name: "Cool DAO"
-    };
-    creatorAddress: "0x1234567890123456789012345678901234567890",
-    metadata: {
-      title: "Test Proposal 2",
-      summary: "Test Proposal Summary 2"
-    };
-    startDate: <Date>,
-    endDate: <Date>,
-    status: "Pending",
-    token: {
-      address: "0x1234567890123456789012345678901234567890,
-      name: "The Token",
-      symbol: "TOK",
-      decimals: 18
-    },
-    results {
-      yes: 100000n,
-      no: 77777n,
-      abstain: 0n
-    }
-  }
-]
+  ]
+}
 ```
 */