+++ modules/client/examples/02-tokenVoting-client/09-get-voting-settings.ts
@@ -9,7 +9,7 @@ import {
   TokenVotingClient,
   VotingSettings
 } from "@aragon/sdk-client";
-import { context } from "../00-setup/00-getting-started";
+import { context } from "../01-client/01-getting-started";
 
 // Instantiate a plugin context from the Aragon OSx SDK context.
 const contextPlugin: ContextPlugin = ContextPlugin.fromContext(context);
@@ -24,13 +24,14 @@ console.log({ pluginSettings });
 /* MARKDOWN
 Returns:
 
-```json
-  {
-    minDuration: 3600, // 1 hour
+```
+{ pluginSettings: {
+    minDuration: 10000, // 10 seconds
     minParticipation: 0.25, // 25%
     supportThreshold: 0.5, // 50%
     minProposerVotingPower: BigInt("5000"),
     votingMode: "Standard"
   }
+}
 ```
 */