+++ modules/client/examples/03-multisig-client/06-can-execute.ts
@@ -8,7 +8,7 @@ import {
   ContextPlugin,
   MultisigClient
 } from "@aragon/sdk-client";
-import { context } from "../00-setup/00-getting-started";
+import { context } from "../01-client/01-getting-started";
 
 // Instantiate a plugin context from the Aragon OSx SDK context.
 const contextPlugin: ContextPlugin = ContextPlugin.fromContext(context);
@@ -24,7 +24,7 @@ console.log({ canExecute });
 /* MARKDOWN
 Returns:
 
-```javascript
+```
   { canExecute: true }
 ```
 */