+++ modules/client/examples/03-multisig-client/08-get-proposal.ts
@@ -9,7 +9,7 @@ import {
   MultisigClient,
   MultisigProposal
 } from "@aragon/sdk-client";
-import { context } from "../00-setup/00-getting-started";
+import { context } from "../01-client/01-getting-started";
 
 // Instantiate a plugin context from the Aragon OSx SDK context.
 const contextPlugin: ContextPlugin = ContextPlugin.fromContext(context);
@@ -24,13 +24,13 @@ console.log({ proposalDetails });
 /* MARKDOWN
 Returns:
 
-```json
+```
 {
   id: "0x1234567890123456789012345678901234567890_0x0",
   dao: {
     address: "0x1234567890123456789012345678901234567890",
     name: "Cool DAO"
-  };
+  },
   creatorAddress: "0x1234567890123456789012345678901234567890",
   metadata: {
     title: "Test Proposal",
@@ -50,7 +50,7 @@ Returns:
       header: "https://.../image.jpeg",
       logo: "https://.../image.jpeg"
     }
-  };
+  },
   creationDate: <Date>,
   actions: [
     {