+++ modules/client/examples/03-multisig-client/09-get-proposals.ts
@@ -13,7 +13,7 @@ import {
   ProposalSortBy,
   ProposalStatus
 } from "@aragon/sdk-client";
-import { context } from "../00-setup/00-getting-started";
+import { context } from "../01-client/01-getting-started";
 
 // Instantiate a plugin context from the Aragon OSx SDK context.
 const contextPlugin: ContextPlugin = ContextPlugin.fromContext(context);
@@ -35,7 +35,7 @@ console.log({ multisigProposals });
 /* MARKDOWN
 Returns:
 
-```json
+```
 [
   {
     id: "0x12345...",