+++ modules/client/examples/05-encoders/15-add-address-list-members.ts
@@ -7,9 +7,9 @@ Adds a list of addresses to the AddressList plugin so that these new addresses a
 import {
   AddresslistVotingClient,
   DaoAction,
-  ContextPlugin }
-from "@aragon/sdk-client";
-import { context } from "../00-setup/00-getting-started";
+  ContextPlugin
+} from "@aragon/sdk-client";
+import { context } from "../01-client/01-getting-started";
 
 // Instantiate a plugin context from the Aragon OSx SDK context
 const contextPlugin: ContextPlugin = ContextPlugin.fromContext(context);