+++ modules/client/examples/05-encoders/18-update-multisig-voting-settings.ts
@@ -12,7 +12,7 @@ import {
   MultisigClient,
   UpdateMultisigVotingSettingsParams
 } from "@aragon/sdk-client";
-import { context } from "../00-setup/00-getting-started";
+import { context } from "../01-client/01-getting-started";
 
 // Instantiate a plugin context from the Aragon OSx SDK context.
 const contextPlugin: ContextPlugin = ContextPlugin.fromContext(context);
@@ -34,7 +34,7 @@ console.log({ updateMultisigConfig });
 /* MARKDOWN
 Returns:
 
-```json
+```
 {
   to: "0x1234567890...",
   value: 0n,