+++ modules/client/examples/06-decoders/11-upgrade-to-action.ts
@@ -5,7 +5,7 @@ Decodes the action of upgrading the DAO to a new implementation.
 */
 
 import { Client, Context, ContextPlugin } from "@aragon/sdk-client";
-import { context } from "../00-setup/00-getting-started";
+import { context } from "../01-client/01-getting-started";
 
 // Initialize the plugin's context from the Aragon SDK context.
 const contextPlugin: ContextPlugin = ContextPlugin.fromContext(context);
@@ -18,7 +18,7 @@ console.log({ upgradeToAction });
 /* MARKDOWN
 Returns:
 
-```json
+```
   { upgradeToAction: "0x1234567890123456789012345678901234567890" }
 ```
 */