+++ modules/client/examples/06-decoders/12-set-signature-validator.ts
@@ -5,7 +5,7 @@ Decodes the action of setting a signature validator for the DAO.
 */
 
 import { Client, ContextPlugin } from "@aragon/sdk-client";
-import { context } from "../00-setup/00-getting-started";
+import { context } from "../01-client/01-getting-started";
 
 // Initialize the plugin's context from the Aragon SDK context.
 const contextPlugin: ContextPlugin = ContextPlugin.fromContext(context);
@@ -18,7 +18,7 @@ console.log({ setSignatureValidatorAction });
 /* MARKDOWN
 Returns:
 
-```json
+```
   { setSignatureValidatorAction: "0x1234567890123456789012345678901234567890" }
 ```
 */