+++ modules/client/examples/06-decoders/15-find-interface-token.ts
@@ -5,7 +5,7 @@ Decodes the parameters of a function call from the TokenVoting plugin contract.
 */
 
 import { ContextPlugin, TokenVotingClient } from "@aragon/sdk-client";
-import { context } from "../00-setup/00-getting-started";
+import { context } from "../01-client/01-getting-started";
 
 // Instantiate a plugin context from the Aragon OSx SDK context.
 const contextPlugin: ContextPlugin = ContextPlugin.fromContext(context);
@@ -21,7 +21,7 @@ console.log({ functionParams });
 /* MARKDOWN
 Returns:
 
-```json
+```
 {
   id: "function functionName(param1, param2)"
   functionName: "functionName"