+++ modules/client/examples/06-decoders/18-find-interface-address-list.ts
@@ -8,7 +8,7 @@ import {
   AddresslistVotingClient,
   ContextPlugin
 } from "@aragon/sdk-client";
-import { context } from "../00-setup/00-getting-started";
+import { context } from "../01-client/01-getting-started";
 
 // Instantiate a plugin context from the Aragon OSx SDK context.
 const contextPlugin: ContextPlugin = ContextPlugin.fromContext(context);
@@ -23,7 +23,7 @@ console.log({ functionParams });
 /* MARKDOWN
 Returns:
 
-```json
+```
 {
   id: "function functionName(param1, param2)"
   functionName: "functionName"