+++ modules/client/examples/06-decoders/20-add-multisig-addresses.ts
@@ -7,7 +7,7 @@ Decodes the parameters of the add members action from the Multisig plugin.
 */
 
 import { ContextPlugin, MultisigClient } from "@aragon/sdk-client";
-import { context } from "../00-setup/00-getting-started";
+import { context } from "../01-client/01-getting-started";
 
 // Instantiate a plugin context from the Aragon OSx SDK context.
 const contextPlugin: ContextPlugin = ContextPlugin.fromContext(context);
@@ -22,11 +22,13 @@ console.log({ decodeAddMembersMultisig });
 /* MARKDOWN
 Returns:
 
-```json
+```
+{ decodeAddMembersMultisig:
   [
     "0x12345...",
     "0x56789...",
     "0x13579..."
   ]
+}
 ```
 */