+++ modules/client/examples/06-decoders/22-update-multisig-voting-settings.ts
@@ -5,7 +5,7 @@ Decodes the update settings action for a Multisig plugin.
 */
 
 import { ContextPlugin, MultisigClient, MultisigVotingSettings } from "@aragon/sdk-client";
-import { context } from "../00-setup/00-getting-started";
+import { context } from "../01-client/01-getting-started";
 
 // Instantiate a plugin context from an Aragon OSx SDK context.
 const contextPlugin: ContextPlugin = ContextPlugin.fromContext(context);
@@ -21,7 +21,7 @@ console.log({ decodeUpdateMultisigSettings });
 /* MARKDOWN
 Returns:
 
-```json
+```
 {
   minApprovals: 2,
   onlyListed: false