+++ modules/common/src/index.ts
@@ -3,7 +3,6 @@ export * from "./encoding";
 export * from "./errors";
 export * from "./promises";
 export * from "./uri";
-export * from "./random";
 export * from "./timeout";
 export * from "./types";
 export * from "./utils";